requirements.txt
alembic==1.8.1
anyio==3.6.1
apache-airflow==2.2.2
apache-airflow-providers-amazon==2.4.0
apache-airflow-providers-common-sql==1.2.0
apache-airflow-providers-ftp==3.1.0
apache-airflow-providers-http==4.0.0
apache-airflow-providers-imap==3.0.0
apache-airflow-providers-sqlite==3.2.1
apispec==3.3.2
argcomplete==1.12.3
asn1crypto==1.5.1
attrs==20.3.0
Babel==2.10.3
beautifulsoup4==4.11.1
blinker==1.5
boto3==1.18.65
botocore==1.21.65
cached-property==1.5.2
cachelib==0.9.0
cattrs==1.5.0
certifi==2022.6.15.1
cffi==1.15.1
charset-normalizer==2.0.12
click==7.1.2
clickclick==20.10.2
colorama==0.4.5
colorlog==5.0.1
commonmark==0.9.1
croniter==1.0.15
cryptography==38.0.1
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.5.1
dnspython==2.2.1
docutils==0.16
email-validator==1.2.1
Flask==1.1.4
Flask-AppBuilder==3.4.5
Flask-Babel==2.0.0
Flask-Caching==1.11.1
Flask-JWT-Extended==3.25.1
Flask-Login==0.4.1
Flask-OpenID==1.3.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.14.3
graphviz==0.20.1
greenlet==1.1.3
gunicorn==20.1.0
h11==0.12.0
httpcore==0.15.0
httpx==0.23.0
idna==3.3
importlib-metadata==4.12.0
importlib-resources==5.9.0
inflection==0.5.1
iso8601==1.0.2
itsdangerous==1.1.0
Jinja2==2.11.3
jmespath==0.10.0
jsonpath-ng==1.5.3
jsonschema==3.2.0
lazy-object-proxy==1.7.1
lockfile==0.12.2
lxml==4.9.1
Mako==1.2.2
Markdown==3.4.1
MarkupSafe==1.1.1
marshmallow==3.17.1
marshmallow-enum==1.5.1
marshmallow-oneofschema==3.0.1
marshmallow-sqlalchemy==0.26.1
mypy-boto3-appflow==1.24.36.post1
mypy-boto3-rds==1.24.61
mypy-boto3-redshift-data==1.24.36.post1
numpy==1.21.6
openapi-schema-validator==0.2.3
openapi-spec-validator==0.4.0
packaging==21.3
pandas==1.3.5
pendulum==2.1.2
ply==3.11
prison==0.2.1
psutil==5.9.2
pycparser==2.21
Pygments==2.13.0
PyJWT==1.7.1
pyparsing==3.0.9
pyrsistent==0.18.1
python-daemon==2.3.1
python-dateutil==2.8.2
python-nvd3==0.15.0
python-slugify==4.0.1
python3-openid==3.2.0
pytz==2022.1
pytzdata==2020.1
PyYAML==6.0
redshift-connector==2.0.908
requests==2.28.0
requests-toolbelt==0.9.1
rfc3986==1.5.0
rich==12.5.1
s3transfer==0.5.2
scramp==1.4.1
setproctitle==1.3.2
six==1.16.0
sniffio==1.3.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.41
SQLAlchemy-JSONField==1.0.0
sqlalchemy-redshift==0.8.11
SQLAlchemy-Utils==0.38.3
sqlparse==0.4.2
swagger-ui-bundle==0.0.9
tabulate==0.8.10
tenacity==8.0.1
termcolor==2.0.0
text-unidecode==1.3
typing_extensions==4.3.0
unicodecsv==0.14.1
urllib3==1.26.12
watchtower==1.0.6
Werkzeug==1.0.1
WTForms==2.3.3
zipp==3.8.1
